synit: mdevd, sysctl, modprobe, fuse, wpa_supplicant, networking

+++ modules/system/activation/default.nix
@@ -113,7 +113,7 @@ in
     system.activation.scripts.specialfs = ''
       echo "specialfs stub here..."
       mkdir -p /bin /etc /run /tmp /usr /var/{cache,db,empty,lib,log,spool}
-      ln -s -n /run/ var/run
+      [ ! -e /var/run ] && ln -s -n /run/ var/run
     '';
 
     system.activation.path = with pkgs; map lib.getBin [
